coolercontrol/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "coolercontrol"
version = "0.14.0"
description = "Monitor and control your cooling and other devices"
authors = ["Guy Boldon <gb@guyboldon.com>"]
license = "GNU General Public License v3.0"
readme = "README.md"
repository = "https://gitlab.com/coolercontrol/coolercontrol"
homepage = "https://gitlab.com/coolercontrol/coolercontrol"
keywords = []
packages = [{ include = "coolercontrol", from = "../" }]

[tool.poetry.urls]
issues = "https://gitlab.com/coolercontrol/coolercontrol/-/issues"

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
pyside6_essentials = "==6.4.1"
apscheduler = "==3.9.1"
matplotlib = "==3.6.2"
numpy = "==1.23.5"
liquidctl = "==1.11.1"
# for dev purposes:
#liquidctl = { git = "https://github.com/liquidctl/liquidctl.git", branch = "main"}
psutil = "==5.9.4"
types-psutil = "==5.9.5.5"
pyamdgpuinfo = "==2.1.3"
setproctitle = "==1.3.2"
jeepney = "==0.8.0"

[tool.poetry.dev-dependencies]
# Nuitka also has the following system dependencies when building:  gcc, ccache, chrpath, patchelf
nuitka = "==1.2.6"
mypy = "==0.991"
pytest = "==7.2.0"
#pytest_mock = "==3.6.1"
#pytest-randomly = "==3.10.3"
#pytest-sugar = "==0.9.4"
#pytest-xdist = "==2.5.0"
#pytest-qt = "==4.0.2"
pylint = "==2.15.8"

[tool.poetry.scripts]
coolercontrol = "coolercontrol:main"
